Fix heading tags: HeadingLevel as u8 to produce <h2> not <hH2>
pulldown-cmark's HeadingLevel enum formats as "H2" with Display,
producing <hH2> tags. Cast to u8 to get the numeric value.

+++ build.rs
@@ -210,10 +210,10 @@ fn render_markdown(content: &str, ss: &SyntaxSet, theme: &syntect::highlighting:
                 html_output.push_str("</code>");
             }
             Event::Start(Tag::Heading { level, .. }) => {
-                html_output.push_str(&format!("<h{}>", level));
+                html_output.push_str(&format!("<h{}>", level as u8));
             }
             Event::End(TagEnd::Heading(level)) => {
-                html_output.push_str(&format!("</h{}>", level));
+                html_output.push_str(&format!("</h{}>", level as u8));
             }
             Event::Start(Tag::Paragraph) => {
                 html_output.push_str("<p>");
